add values for alarm data, add count for failsafe

+++ packs/Nissan/LeafPack.cpp
@@ -26,6 +26,7 @@ LeafPack::LeafPack(
    m_reboot_in_process(false),
    m_shunt_fail_msg_logged(false),
    m_reboot_wait_count(0),
+   m_failsafe_count(0),
    m_log(log)
 {
    std::string pcname; pcname.append(m_pack_name).append("_LeafPackPeriodic");
@@ -88,12 +89,22 @@ void LeafPack::heartbeatCallback()
                && m_reboot_wait_count > REBOOT_WAIT_PERIODS
                && m_monitor.getPackStatus() == monitor::Monitor::NORMAL_OPERATION)
          {
-            m_reboot_wait_count = 0;
-            m_reboot_in_process = true;
-            std::ostringstream ss;
-            ss << "LeafPack: " << m_pack_name << ": Failsafe Status indicates Pack needs a reboot";
-            if (m_log) m_log->alarm(ss, __FILENAME__, __LINE__);
-            m_power_relay.setState(contactor::Nissan::TeensyRelay::ENERGIZED);
+            // failsafe has to be on for FAILSAFE_COUNT consecutive periods
+            m_failsafe_count++;
+            std::ostringstream sss;
+            sss << "LeafPack: " << m_pack_name << ": Failsafe Status indicator on: count = " << m_failsafe_count;
+            if (m_log) m_log->alarm(sss, __FILENAME__, __LINE__);
+            if (m_failsafe_count > FAILSAFE_COUNT)
+            {
+               // reboot
+               m_failsafe_count = 0;
+               m_reboot_wait_count = 0;
+               m_reboot_in_process = true;
+               std::ostringstream ss;
+               ss << "LeafPack: " << m_pack_name << ": Failsafe Status indicates Pack needs a reboot";
+               if (m_log) m_log->alarm(ss, __FILENAME__, __LINE__);
+               m_power_relay.setState(contactor::Nissan::TeensyRelay::ENERGIZED);
+            }            
          }
          else if (m_reboot_in_process && (m_reboot_wait_count > REBOOT_POWERDOWN_PERIODS))
          {
@@ -103,6 +114,10 @@ void LeafPack::heartbeatCallback()
                   << REBOOT_WAIT_PERIODS * PACK_CALLBACK_PERIOD_ms / 1000 << " seconds";
             if (m_log) m_log->alarm(ss, __FILENAME__, __LINE__);
             m_power_relay.setState(contactor::Nissan::TeensyRelay::DE_ENERGIZED);
+         }
+         else
+         {
+            m_failsafe_count = 0;
          }
          break;
